+++ src/functions/checks.ts
@@ -1,4 +1,6 @@
-import { PREFIX } from '../data/config';
+import fs from 'fs';
+import path from 'path';
+import { CONNECTION_SAVE_PATH, PREFIX, TEMP_FOLDER_PATH } from '../data/config';
 import { getOwnerJid, JID_GROUP_SUFIX } from './jids-funcs';
 
 export const checkIsCommand = (message: string) => {
@@ -13,4 +15,19 @@ export const checkIsGroup = (jid: string) => {
 export const checkIsWoner = (jid: string) => {
     const wonerJid = getOwnerJid();
     return wonerJid === jid;
+}
+
+export const checksDependenciesDirs = () => {
+    const hasTempFolder = fs.existsSync(TEMP_FOLDER_PATH);
+    const hasConnectionFolder = fs.existsSync(CONNECTION_SAVE_PATH);
+
+    if(!hasTempFolder) {
+        fs.mkdirSync(TEMP_FOLDER_PATH);
+        console.log('   > TEMP-FOLDER criado em '+ TEMP_FOLDER_PATH);
+    }
+    if(!hasConnectionFolder) {
+        fs.mkdirSync(CONNECTION_SAVE_PATH);
+        console.log('   > CONNECT SAVE FOLDER criado em '+ CONNECTION_SAVE_PATH);
+    }
+
 }
